Fixed having less than the required number of bits for the data length

+++ demboyz/netmessages/svc_gameeventlist.h
@@ -21,7 +21,7 @@ namespace NetMsg
             std::vector<EventValue> values;
         };
 
-        uint16_t dataLengthInBits;
+        uint32_t dataLengthInBits;
         std::vector<EventDescriptor> eventDescriptors;
     };
 }
